tools/requirements.txt
azure-storage-blob>=12.0.0